Implement the Whisper engine directly on ONNX Runtime
Transcribes real speech correctly, GPU-accelerated, no wrapper library
between the app and the model.

    audio:   11.0s JFK sample
    CPU  int8  8.73s  1.26x realtime
    CUDA int8  4.06s  2.71x realtime
    "And so, my fellow Americans, ask not what your country can do for
     you, ask what you can do for your country."

mel.py is the only real DSP: a Whisper-exact log-mel front-end in numpy,
no torch or librosa. The details that look incidental are the ones that
matter -- centered STFT, periodic Hann window, the dropped final frame
that makes 30s land on exactly 3000 frames, and Slaney (not HTK) mel
normalization. A wrong front-end here does not raise, it makes the model
hallucinate fluently, so the tests pin those invariants rather than
output text.

engine.py drives the two graphs directly. The export makes this
practical: the KV cache is an explicit input/output pair plus an offset,
so generation is linear rather than quadratic in output length, and the
encoder emits cross-attention K/V directly so it runs once per 30s
window. Every Whisper constant is read from the model metadata rather
than hardcoded.

Three things the spike changed:

- The plan was fp16 on GPU, on the theory that ORT's CUDA EP handles
  quantized ops badly enough that int8-on-CUDA would lose to CPU. That
  was wrong: int8 on CUDA is 2.15x faster than CPU. One int8 model now
  serves every backend, halving the first-run download.

- ONNX Runtime falls back to CPU *silently* when a provider's native
  library is missing -- correct output, half the speed, one stderr
  warning. get_available_providers() reports what ORT was built with,
  not what it can load, so backend.confirm() re-reads the truth from
  session.get_providers() after construction.

- onnxruntime-gpu does not bundle the CUDA runtime. For CUDA 13 those
  wheels live on pypi.nvidia.com rather than PyPI, and they use a new
  nvidia/cu13/lib layout that ORT 1.30 does not search, so cuda.py
  dlopens them with RTLD_GLOBAL before any session is created. Without
  it CUDA never engages and nothing says so.

audio.py and the app loop are written but unproven: libportaudio2 is not
installed here, so sounddevice fails at import. Note that PulseAudio
exposes "Monitor of ..." loopback sources alongside real microphones --
picking one silently records the speakers, so they are filtered out.

+++ tests/test_backend.py
@@ -11,7 +11,7 @@ def test_cpu_only_falls_back_to_int8():
 def test_cuda_selects_fp16_and_keeps_cpu_as_fallback():
     be = resolve(["CUDAExecutionProvider", "CPUExecutionProvider"])
     assert be.providers == ["CUDAExecutionProvider", "CPUExecutionProvider"]
-    assert be.quantization == "fp16"
+    assert be.quantization == "int8"
     assert be.label == "CUDA"
     assert be.accelerated
 
@@ -19,7 +19,7 @@ def test_cuda_selects_fp16_and_keeps_cpu_as_fallback():
 def test_coreml_is_picked_on_macos():
     be = resolve(["CoreMLExecutionProvider", "CPUExecutionProvider"])
     assert be.label == "CoreML"
-    assert be.quantization == "fp16"
+    assert be.quantization == "int8"
 
 
 def test_cuda_wins_over_other_accelerators():
